requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=4.5.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
gunicorn>=21.0.0
django-celery-beat>=2.5.0

